+++ src/Conditions/HasCompletedEndingCondition.cs
@@ -1,22 +1,20 @@
 using BingoAPI.Conditions;
+using Newtonsoft.Json;
 using Silksong.BingoSync.Data;
 using Silksong.BingoSync.Extensions;
 
 namespace Silksong.BingoSync.Conditions;
 
 /// <summary>
-/// Checks if the player has completed a given <see cref="Ending"/>
+/// Checks if the player has completed a given <see cref="Data.Ending"/>
 /// </summary>
+[Condition("has_completed_ending")]
 internal sealed class HasCompletedEndingCondition : ICondition
 {
-	private readonly Ending _ending;
-
-	[Condition("has_completed_ending")]
-	public HasCompletedEndingCondition(ConditionData data)
-	{
-		_ending = data.GetRequiredParameter<Ending>("ending");
-	}
+	[JsonProperty("ending")]
+	[JsonRequired]
+	public required Ending Ending { get; init; }
 
 	/// <inheritdoc />
-	public bool IsMet() => PlayerData.instance.IsEndingCompleted(_ending);
+	public bool IsMet() => PlayerData.instance.HasCompletedEnding(Ending);
 }