+++ src/Conditions/HasObtainedSilkSkillCondition.cs
@@ -1,4 +1,5 @@
 using BingoAPI.Conditions;
+using Newtonsoft.Json;
 using Silksong.BingoSync.Data;
 using Silksong.BingoSync.Extensions;
 
@@ -7,16 +8,13 @@ namespace Silksong.BingoSync.Conditions;
 /// <summary>
 /// Checks if the player has obtained a given <see cref="Crest"/>
 /// </summary>
+[Condition("has_obtained_silk_skill")]
 internal sealed class HasObtainedSilkSkillCondition : ICondition
 {
-	private readonly SilkSkill _skill;
-
-	[Condition("has_obtained_silk_skill")]
-	public HasObtainedSilkSkillCondition(ConditionData data)
-	{
-		_skill = data.GetRequiredParameter<SilkSkill>("skill");
-	}
+	[JsonProperty("skill")]
+	[JsonRequired]
+	public required SilkSkill Skill { get; init; }
 
 	/// <inheritdoc />
-	public bool IsMet() => PlayerData.instance.HasSilkSkill(_skill);
+	public bool IsMet() => PlayerData.instance.HasObtainedSilkSkill(Skill);
 }