+++ src/Conditions/HasObtainedToolCondition.cs
@@ -1,22 +1,20 @@
 using BingoAPI.Conditions;
+using Newtonsoft.Json;
 using Silksong.BingoSync.Data;
 using Silksong.BingoSync.Extensions;
 
 namespace Silksong.BingoSync.Conditions;
 
 /// <summary>
-/// Checks if the player has obtained a given <see cref="Tool"/>
+/// Checks if the player has obtained a given <see cref="Data.Tool"/>
 /// </summary>
+[Condition("has_obtained_tool")]
 internal sealed class HasObtainedToolCondition : ICondition
 {
-	private readonly Tool _tool;
-
-	[Condition("has_obtained_tool")]
-	public HasObtainedToolCondition(ConditionData data)
-	{
-		_tool = data.GetRequiredParameter<Tool>("tool");
-	}
+	[JsonProperty("tool")]
+	[JsonRequired]
+	public required Tool Tool { get; init; }
 
 	/// <inheritdoc />
-	public bool IsMet() => PlayerData.instance.HasTool(_tool);
+	public bool IsMet() => PlayerData.instance.HasObtainedTool(Tool);
 }